+++ ydb/core/mon/ut_utils/ut_utils.cpp
@@ -7,6 +7,12 @@ namespace NMonitoring::NTests {
 using namespace NActors;
 using namespace NKikimr;
 
+const TString TEST_MON_PATH = "test_mon";
+const TString TEST_RESPONSE = "Test actor";
+const TString AUTHORIZATION_HEADER = "Authorization";
+const TString VALID_TOKEN = "Bearer token";
+const TVector<TString> DEFAULT_TICKET_PARSER_GROUPS = {"group_name"};
+
 void TTestActorPage::Bootstrap() {
     Become(&TTestActorPage::StateWork);
 }
