+++ ydb/core/mon/ya.make
@@ -27,6 +27,11 @@ PEERDIR(
 
 END()
 
+RECURSE(
+    audit
+    ut_utils
+)
+
 RECURSE_FOR_TESTS(
     ut
 )
