+++ ydb/core/viewer/ut/ya.make
@@ -17,7 +17,7 @@ SRCS(
 
 PEERDIR(
     ydb/core/mon
-    ydb/core/mon/ut/ut_utils
+    ydb/core/mon/ut_utils
     library/cpp/http/misc
     library/cpp/http/simple
     ydb/core/testlib/default
